+++ docs/release-notes.md
@@ -3,6 +3,7 @@
 ### [x.x.x]
 - `voyager login` now asks for the artifactory url and username
 - Remove artifactory_url default value from config templates
+- CMake and MSBuild generators now create a variable pointing to the `.voyager` directory.
 
 ### [1.15.0]
 - Add `voyager login` so no more copying of API key is needed